requirements.txt
sounddevice==0.4.5
requests==2.28.2
websocket-client==1.5.0
urllib3==1.26.14
selenium==4.8.0


# sudo apt install -y chromium-browser
    #  sudo apt-add-repository universe
    # sudo apt-get update
    # sudo apt-get install -y ubuntu-restricted-extras
    # sudo apt-get update
        # sudo snap install chromium
# sudo apt-get install -y libportaudio2